EH Location is now modifyable, site management page is active for admins, reportable workouts, clean up

+++ src/app/pages/user-detail/user-detail.component.ts
@@ -1,8 +1,11 @@
 import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Observable, Subject } from 'rxjs';
+import { LinkSiteQAODialog } from 'src/app/dialogs/link-site-q-ao/link-site-q-ao-dialog.component';
 import { AuthenticatedUser, UserRole } from 'src/app/models/authenticated-user.model';
 import { IPaxUser } from 'src/app/models/users.model';
+import { AOManagerService } from 'src/app/services/ao-manager.service';
 import { PaxManagerService } from 'src/app/services/pax-manager.service';
 import { UserAuthenticationService } from 'src/app/services/user-authentication.service';
 
@@ -19,11 +22,15 @@ export class UserDetailComponent {
   public isAdmin = false; // Only admins can promote to admin
   public isAuthorizedUser = false; // Can view content and promote to SiteQ
   public editMode: boolean = false;
+  adminErrorMessage: string = "";
 
   constructor(
     private readonly paxManagerService: PaxManagerService,
     private activatedRoute: ActivatedRoute,
-    private userAuthService: UserAuthenticationService
+    private userAuthService: UserAuthenticationService,
+    private dialog: MatDialog,
+    private readonly aoManagerService: AOManagerService,
+    private router: Router
   ) {
     // Look at the logged in user for admin / siteq permissions
     this.userAuthService.authUserData$.subscribe((res) => {
@@ -58,8 +65,46 @@ export class UserDetailComponent {
 
   public async promoteRole(role: string, user: IPaxUser) {
     const userRole: UserRole = UserRole[role as keyof typeof UserRole];
-    await this.userAuthService.promoteRole(userRole, user);
+    try {
+      const succeeds = await this.userAuthService.validatePromoteRole(user);
+      if (succeeds) {
+        if (userRole === UserRole.SiteQ) {
+          const success = await this.validateSiteQLocation(user);
+          if (!success) {
+            return;
+          }
+        }
+        await this.userAuthService.promoteRole(userRole, user);
+      } else {
+        return;
+      }
+    } catch(err) {
+      if (err instanceof Error) {
+        this.adminErrorMessage = err.message;
+      }
+      return;
+    }
     this.existingRoles.push(userRole);
+  }
+
+  public async validateSiteQLocation(user: IPaxUser): Promise<boolean> {
+    const dialogRef = this.dialog.open(LinkSiteQAODialog);
+    return dialogRef.afterClosed().toPromise().then(async (data) => {
+      if (data && data.aoRef !== '') {
+        await this.linkSiteQAndAO(data.aoRef, user);
+        return true;
+      } else {
+        return false;
+      }
+    });
+  }
+
+  public async linkSiteQAndAO(locationDbPath: string, user: IPaxUser) {
+    const aoRef = this.aoManagerService.getAoLocationReference(locationDbPath);
+    const userRef = this.paxManagerService.getUserReference(`users/${user.id}`);
+    if (aoRef && userRef) {
+      await this.aoManagerService.updateSiteQUser(aoRef, userRef);
+    }
   }
 
   public async getCurrentUserRoles(userId: string) {
@@ -74,6 +119,10 @@ export class UserDetailComponent {
   }
 
   public async deleteUser(user: IPaxUser): Promise<void> {
-    return this.paxManagerService.deleteUser(user);
+    if (confirm("Are you sure you want to *permanantly* remove this Pax data?")) {
+      await this.paxManagerService.deleteUser(user);
+      alert("User deleted");
+      await this.router.navigate(['search']);
+    }
   }
 }
